# shared-contracts/target-operational-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "CursorPageInfo": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "maxLength": 1024,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "returnedCount": {
            "minimum": 0.0,
            "title": "Returnedcount",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "returnedCount",
          "hasMore",
          "nextCursor"
        ],
        "title": "CursorPageInfo",
        "type": "object"
      },
...
      "EnterpriseNotificationCreate": {
        "properties": {
          "enterpriseId": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Enterpriseid",
            "type": "string"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "default": "Warning",
            "enum": [
              "Info",
              "Warning",
              "Critical",
              "Success"
            ],
            "title": "Severity",
            "type": "string"
          },
          "sourceId": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          },
          "sourceType": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetype"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "Staff Follow-up",
            "maxLength": 60,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "enterpriseId",
          "title",
          "message"
        ],
        "title": "EnterpriseNotificationCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationReadUpdate": {
        "properties": {
          "read": {
            "default": true,
            "title": "Read",
            "type": "boolean"
          }
        },
        "title": "NotificationReadUpdate",
        "type": "object"
      },
...
      "UserNotificationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserNotificationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/CursorPageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "UserNotificationPage",
        "type": "object"
      },
      "UserNotificationSummary": {
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readat"
          },
          "recipientAccountId": {
            "title": "Recipientaccountid",
            "type": "string"
          },
          "recipientEnterpriseId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipiententerpriseid"
          },
          "recipientRole": {
            "title": "Recipientrole",
            "type": "string"
          },
          "severity": {
            "enum": [
              "Info",
              "Warning",
              "Critical",
              "Success"
            ],
            "title": "Severity",
            "type": "string"
          },
          "sourceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          },
          "sourceType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetype"
          },
          "targetPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetpath"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "recipientAccountId",
          "title",
          "message",
          "type",
          "severity",
          "recipientRole",
          "createdAt"
        ],
        "title": "UserNotificationSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TANAW API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operational/notifications": {
      "get": {
        "operationId": "list_notifications_operational_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotificationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/operational/notifications/enterprise": {
      "post": {
        "operationId": "create_enterprise_notification_operational_notifications_enterprise_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseNotificationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotificationSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Enterprise Notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/operational/notifications/{notification_id}": {
      "patch": {
        "operationId": "update_notification_read_status_operational_notifications__notification_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationReadUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotificationSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Notification Read Status",
        "tags": [
          "notifications"
        ]
      }
    },
...
  }
}